+++ packages/clap/bili.config.js
@@ -2,5 +2,5 @@ module.exports = {
   filename: 'clap.bin.js',
   input: 'cli.js',
   outDir: 'bin/',
-  format: ['cjs']
+  format: ['cjs'],
 };